requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
antlr4-python3-runtime==4.11.1
anyio==4.11.0
attrs==25.4.0
beautifulsoup4==4.14.2
bs4==0.0.2
certifi==2025.10.5
cfgv==3.4.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docstring_parser==0.17.0
dotenv==0.9.9
et_xmlfile==2.0.0
fastapi==0.120.4
filelock==3.20.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hydra-core==1.4.0.dev1
identify==2.6.15
idna==3.11
iniconfig==2.3.0
instructor==1.12.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==1.0.3
langchain-core==1.0.2
langchain-deepseek==1.0.0
langchain-ollama==1.0.0
langchain-openai==1.0.1
langgraph==1.0.2
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langserve==0.3.3
langsmith==0.4.39
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
networkx==3.5
nodeenv==1.9.1
numpy==2.3.4
ollama==0.6.0
omegaconf==2.4.0.dev3
openai==1.109.1
openpyxl==3.1.5
orjson==3.11.4
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
pre_commit==4.3.0
propcache==0.4.1
pyarrow==22.0.0
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
seaborn==0.13.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.49.3
swarm @ git+https://github.com/openai/swarm.git@0c82d7d868bb8e2d380dfd2a319b5c3a1f4c0cb9
sympy==1.13.1
tenacity==9.1.2
tiktoken==0.12.0
timeout-decorator==0.5.0
torch==2.9.0
tqdm==4.67.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
virtualenv==20.35.3
wheel==0.45.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
